Agregado acción del presupuesto

+++ src/main/java/com/mycompany/proyecto2p/Ciudad.java
@@ -6,10 +6,10 @@ public class Ciudad {
     private String nombre;
     private String nombreAlcalde;
     private Tiempo tiempo;
-    private int presupuesto;
+    private double presupuesto;
     private ArrayList <Servicio> servicios=new ArrayList<>();
 
-    public Ciudad(String nombre, String nombreAlcalde, int presupuesto) {
+    public Ciudad(String nombre, String nombreAlcalde, double presupuesto) {
         this.nombre = nombre;
         this.nombreAlcalde = nombreAlcalde;
         this.presupuesto = presupuesto;
@@ -27,6 +27,7 @@ public class Ciudad {
         this.presupuesto = presupuesto;
     }
     
+    //Metodos que sirven para cuando se gaste o gane dinero en la ciudad.
     public void disminuirPresupuesto(int cantidad){
         presupuesto=presupuesto-cantidad;
     }
